+++ Killboard.Data/Models/corporations.cs
@@ -15,11 +15,11 @@ namespace Killboard.Data.Models
 
         public string name { get; set; }
         public int? alliance_id { get; set; }
-        public int? ceo_char_id { get; set; }
+        public long? ceo_char_id { get; set; }
         public string description { get; set; }
         public string ticker { get; set; }
         public DateTime create_time { get; set; }
-        public int? create_char { get; set; }
+        public long? create_char { get; set; }
         public int member_count { get; set; }
         public decimal? tax_rate { get; set; }
         public string url { get; set; }
