+++ Killboard.Domain/DTO/Character/GetCharacter.cs
@@ -6,7 +6,7 @@ namespace Killboard.Domain.DTO.Character
 {
     public class GetCharacter
     {
-        public int CharacterID{ get; set; }
+        public long CharacterID { get; set; }
         public string Username { get; set; }
         public int CorporationID { get; set; }
         public int? AllianceID { get; set; }
